Moving DataState into core models.
