--- data/core/README.md
@@ -1,3 +1,0 @@
-# Core Data
-
-This module contains any classes that should be shared among any of the various data layers in this application. For example, we use [DataState](src/commonMain/kotlin/com/adammcneilly/pocketleague/core/data/DataState.kt) which encapsulates the possible result types of a data request.
