--- data/core/build.gradle.kts
@@ -1,53 +1,0 @@
-plugins {
-    kotlin("multiplatform")
-}
-
-kotlin {
-    jvm()
-
-    sourceSets {
-        val commonMain by getting
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-        maybeCreate("iosX64Main")
-        maybeCreate("iosArm64Main")
-        maybeCreate("iosSimulatorArm64Main")
-        maybeCreate("iosMain").apply {
-            dependsOn(commonMain)
-            getAt("iosX64Main").dependsOn(this)
-            getAt("iosArm64Main").dependsOn(this)
-            getAt("iosSimulatorArm64Main").dependsOn(this)
-        }
-        maybeCreate("iosX64Test")
-        maybeCreate("iosArm64Test")
-        maybeCreate("iosSimulatorArm64Test")
-        maybeCreate("iosTest").apply {
-            dependsOn(commonTest)
-            getAt("iosX64Test").dependsOn(this)
-            getAt("iosArm64Test").dependsOn(this)
-            getAt("iosSimulatorArm64Test").dependsOn(this)
-        }
-    }
-}
-
-project.extensions.findByType(org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension::class.java)?.apply {
-    if (project.findProperty("ios") == "true") {
-        listOf(
-            iosX64(),
-            iosArm64(),
-            iosSimulatorArm64()
-        ).forEach {
-            it.binaries.framework {
-                baseName = project.name
-            }
-        }
-    }
-    if (project.findProperty("js") == "true") {
-        js(IR) {
-            browser()
-        }
-    }
-}
