--- data/core/src/commonTest/kotlin/com/adammcneilly/pocketleague/core/data/DataStateTest.kt
@@ -1,61 +1,0 @@
-package com.adammcneilly.pocketleague.core.data
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class DataStateTest {
-
-    @Test
-    fun mapLoading() {
-        val loading: DataState<Int> = DataState.Loading
-        val mapped: DataState<String> = loading.map {
-            ""
-        }
-
-        assertEquals(
-            expected = DataState.Loading,
-            actual = mapped,
-        )
-    }
-
-    @Test
-    fun mapError() {
-        val error: DataState<Int> = DataState.Error(Throwable("Whoops"))
-        val mapped: DataState<String> = error.map {
-            ""
-        }
-
-        assertEquals(
-            expected = "Whoops",
-            actual = (mapped as DataState.Error).error.message,
-        )
-    }
-
-    @Test
-    fun mapSuccess() {
-        val success: DataState<Int> = DataState.Success(1)
-        val mapped: DataState<String> = success.map {
-            it.toString()
-        }
-
-        assertEquals(
-            expected = DataState.Success("1"),
-            actual = mapped,
-        )
-    }
-
-    @Test
-    fun errorWhenMapping() {
-        val error = IllegalArgumentException("Invalid mapping.")
-
-        val success: DataState<Int> = DataState.Success(1)
-        val mapped: DataState<String> = success.map {
-            throw error
-        }
-
-        assertEquals(
-            expected = DataState.Error(error),
-            actual = mapped,
-        )
-    }
-}
