+++ data/event-test/build.gradle.kts
@@ -11,7 +11,6 @@ kotlin {
             dependencies {
                 implementation(project(":core:models"))
                 implementation(project(":core:models-test"))
-                implementation(project(":data:core"))
                 implementation(project(":data:event"))
             }
         }
