+++ data/event-test/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/event/test/FakeEventService.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.data.event.test
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.models.Event
 import com.adammcneilly.pocketleague.core.models.Team
 import com.adammcneilly.pocketleague.core.models.test.TestModel
