+++ data/event/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/event/EventService.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.data.event
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.models.Event
 import com.adammcneilly.pocketleague.core.models.Team
