+++ data/game/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/game/GameService.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.data.game
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.models.Game
 
 /**
