+++ data/match-test/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/match/test/FakeMatchService.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.data.match.test
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.models.Match
 import com.adammcneilly.pocketleague.core.models.test.TestModel
 import com.adammcneilly.pocketleague.core.models.test.matchBlueWinner
