+++ data/match/src/commonMain/kotlin/com/adammcneilly/pocketleague/data/match/MatchService.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.data.match
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.models.Match
 import kotlinx.coroutines.flow.Flow
