+++ data/octanegg/build.gradle.kts
@@ -11,7 +11,6 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(project(":core:models"))
-                implementation(project(":data:core"))
                 implementation(project(":data:remote"))
                 implementation(libs.kotlinx.serialization.json)
             }
