+++ shared/build.gradle.kts
@@ -13,7 +13,6 @@ kotlin {
                 implementation(project(":core:datetime"))
                 implementation(project(":core:displaymodels"))
                 implementation(project(":core:models"))
-                implementation(project(":data:core"))
                 implementation(project(":data:event"))
                 implementation(project(":data:game"))
                 implementation(project(":data:match"))
