+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/screens/myteams/LoadMyTeamsEvent.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.shared.screens.myteams
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.displaymodels.toDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.toOverviewDisplayModel
 import com.adammcneilly.pocketleague.core.models.Match
