+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/screens/myteams/MyTeamsViewState.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.shared.screens.myteams
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.TeamOverviewDisplayModel
 import com.adammcneilly.pocketleague.shared.screens.ScreenState
