+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/screens/teamselection/LoadTeamSelectionEvent.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.shared.screens.teamselection
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.displaymodels.toOverviewDisplayModel
 import com.adammcneilly.pocketleague.core.models.Team
 import com.adammcneilly.pocketleague.shared.screens.Events
