+++ shared/src/commonTest/kotlin/com/adammcneilly/pocketleague/shared/screens/matchdetail/MatchDetailViewStateTest.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.shared.screens.matchdetail
 
-import com.adammcneilly.pocketleague.core.data.DataState
+import com.adammcneilly.pocketleague.core.models.DataState
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.test.TestDisplayModel
 import kotlin.test.Test
